unit05: testing

+++ docs/unit05/scripts/groceries.py
@@ -17,11 +17,11 @@ def compute_average_quantity(a_list_of_dicts, a_key_string)-> float:
         result (float): Average value.
     """
 
-    total_quantity = 0
+    total_quantity = 0.
     for item in a_list_of_dicts:
-        total_quantity += item[ a_key_string ]
+        total_quantity += float(item[ a_key_string ])
 
-    return int( total_quantity / len( a_list_of_dicts ) )
+    return ( total_quantity / len( a_list_of_dicts ) )
 
 def calc_total_price( price, quantity )-> float:
     """
